# api/openapi.yaml
openapi: 3.0.3
info:
  title: Host API
  version: 0.1.0
  description: API for experiment hosting platform with jsPsych, DataPipe, and Prolific integration
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.yoursite.com
    description: Production server
paths:
...
  /run/{id}:
    get:
      tags: [Public]
      summary: Run experiment (public endpoint for participants)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: PROLIFIC_PID
          schema:
            type: string
          description: Prolific participant ID
      responses:
        '200':
          description: Experiment runner HTML page
          content:
            text/html:
              schema:
                type: string
        '403':
          description: Experiment not live
        '404':
          description: Experiment not found
  /run/{id}/data:
    post:
      tags: [Public]
      summary: Submit experiment data (public endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                prolific_pid:
                  type: string
                  nullable: true
                data:
                  type: object
              required: [session_id, data]
      responses:
        '200':
          description: Data saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string, format: uuid }
                  message: { type: string }
  /run/{id}/complete:
    get:
      tags: [Public]
      summary: Experiment completion page (public endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Completion page HTML
          content:
            text/html:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
...